✨ 重命名插件名称；
