+++ package.json
@@ -8,7 +8,7 @@
     "build": "esno scripts/build.ts",
     "publish": "esno scripts/publish.ts",
     "compress": "esno scripts/compress.ts",
-    "clean": "rimraf dcloud dist"
+    "clean": "rimraf dcloud dist template/js_sdk"
   },
   "files": [
     "dist",
