+++ packages/plugins/drawImage/index.ts
@@ -65,4 +65,4 @@ const plugin: DrawPosterPlugin = {
   }
 }
 
-export default plugin
+export default () => plugin
