+++ packages/plugins/drawImage/utils.ts
