+++ packages/plugins/drawImageFit/index.ts
@@ -55,4 +55,4 @@ const plugin: DrawPosterPlugin = {
   }
 }
 
-export default plugin
+export default () => plugin
