+++ packages/plugins/drawImageFit/object-sizing.ts
