+++ packages/plugins/drawQrCode/uQRCode.d.ts
