+++ packages/plugins/drawQrCode/uQRCode.js
