+++ packages/plugins/drawRoundImage/index.ts
@@ -31,4 +31,4 @@ const plugin: DrawPosterPlugin = {
   }
 }
 
-export default plugin
+export default () => plugin
