+++ packages/plugins/fillWarpText/index.ts
@@ -97,4 +97,4 @@ const plugin: DrawPosterPlugin = {
   }
 }
 
-export default plugin
+export default () => plugin
