+++ packages/plugins/gcanvas/gcanvas/bridge/bridge-weex.js
