+++ packages/plugins/gcanvas/gcanvas/context-2d/FillStyleLinearGradient.js
