+++ packages/plugins/gcanvas/gcanvas/context-2d/FillStylePattern.js
