+++ packages/plugins/gcanvas/gcanvas/context-2d/FillStyleRadialGradient.js
