+++ packages/plugins/gcanvas/gcanvas/context-webgl/ActiveInfo.js
