+++ packages/plugins/gcanvas/gcanvas/context-webgl/ShaderPrecisionFormat.js
