+++ packages/plugins/gcanvas/gcanvas/context-webgl/classUtils.js
