+++ packages/plugins/gcanvas/gcanvas/index.d.ts
