+++ packages/plugins/gcanvas/gcanvas/index.js
