+++ packages/plugins/gcanvas/index.ts
@@ -14,4 +14,4 @@ const plugin: DrawPosterPlugin = {
   }
 }
 
-export default plugin
+export default () => plugin
