+++ packages/plugins/index.ts
@@ -1,9 +1,9 @@
-export { default as drawImageFit } from './ctx-drawImageFit'
-export { default as drawQrCode } from './ctx-drawQrCode'
-export { default as drawRoundImage } from './ctx-drawRoundImage'
-export { default as fillRoundRect } from './ctx-fillRoundRect'
-export { default as fillWarpText } from './ctx-fillWarpText'
-export { default as roundRect } from './ctx-roundRect'
-export { default as strokeRoundRect } from './ctx-strokeRoundRect'
-export { default as painter } from './dp-painter'
-export { default as table } from './dp-table'
+export { default as drawImageFit } from './drawImageFit'
+export { default as drawQrCode } from './drawQrCode'
+export { default as drawRoundImage } from './drawRoundImage'
+export { default as fillRoundRect } from './fillRoundRect'
+export { default as fillWarpText } from './fillWarpText'
+export { default as roundRect } from './roundRect'
+export { default as strokeRoundRect } from './strokeRoundRect'
+export { default as painter } from './painter'
+export { default as table } from './table'
