+++ packages/plugins/roundRect/index.ts
@@ -57,4 +57,4 @@ const plugin: DrawPosterPlugin = {
   }
 }
 
-export default plugin
+export default () => plugin
