+++ template/manifest.json
@@ -52,7 +52,8 @@
     "mp-weixin" : {
         "appid" : "",
         "setting" : {
-            "urlCheck" : false
+            "urlCheck" : false,
+            "checkSiteMap": false
         },
         "usingComponents" : true
     },
